somehow the code fixed itself, no more unrecognized imag parts. Added config option to take abs value of all mass squares

+++ src/ThreeHiggs/parsedmatrix.py
@@ -123,5 +123,5 @@ class ParsedMatrix:
         # should not have any free symbols
         assert len(sympyMatrix.free_symbols) == 0, "Failed to parse constant matrix: has free symbols!"
         # can't lambdify with no arguments, so do standard conversion to numpy array instead
-        return np.array(sympyMatrix.tolist())
+        return np.array(sympyMatrix.tolist()).astype(np.float64)
